:sparkles: feat: Implement PUT API for member after signup to set info

+++ src/main/kotlin/com/team1/mvp_test/domain/member/dto/MemberResponse.kt
@@ -6,7 +6,7 @@ data class MemberResponse(
     val id: Long,
     val name: String?,
     val email: String?,
-    val age: Long?,
+    val age: Int?,
     val sex: String?,
     val info: String?,
     val signupState: Boolean,
