+++ src/main/kotlin/com/team1/mvp_test/domain/member/model/Member.kt
@@ -18,7 +18,7 @@ class Member(
     val email: String,
 
     @Column(name = "age")
-    var age: Long? = null,
+    var age: Int? = null,
 
     @Column(name = "sex")
     var sex: String? = null,
